+++ .gitignore
@@ -2,3 +2,7 @@ node_modules/
 .env
 nohup.out
 logfile
+/redis-stable
+redis-stable.tar.gz
+tester.psql
+.DS_Store
