+++ config.js
@@ -1,0 +1,3 @@
+var config = { // this should be const instead if you're using ES6 standards
+  apiKey : 'https://maps.googleapis.com/maps/api/js?key=AIzaSyBt2uGZOui5nESr2PwoRVJxP2OUqYA-gp8&callback=initMap'
+}
