+++ routes/index.js
@@ -4,11 +4,14 @@
 */
 const coins = require('./coins')
 const users = require('./users')
+const admin = require('./admin')
 
 //put routes inside this exports module 
 module.exports = (app) => {
     app.use('/coins', coins)
     // just syntax to use what's in the "coins" route 
-    app.use('/users', users) // <<---- HERE!
+    app.use('/users', users) 
+    app.use('/admin', admin) 
+
 }
