+++ views/logindex.ejs
@@ -1,6 +1,6 @@
 <h1>Login</h1>
 
-<form action=" http://localhost:3000/loginsession" method = "POST">
+<form action=" http://localhost:3000/users/loginsession" method = "POST">
     <div>
         <label for = "email"> Email</label> 
         <input type="email" id = "email" name = "email"
